Adds read functionality and updates builtins
Adds the `read` functionality to the shell.
Updates the `is_builtin` function to use a more readable if-else structure.
Adds test.c to the list of source files.
