+++ includes/libft/srcs/ft_strcmp.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strcmp.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: malrifai <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: eaqrabaw <eaqrabaw@student.42amman.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/20 12:40:44 by malrifai          #+#    #+#             */
-/*   Updated: 2024/11/10 00:04:41 by malrifai         ###   ########.fr       */
+/*   Updated: 2025/04/07 06:45:09 by eaqrabaw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
